+++ simbench/networks/extract_simbench_grids_from_csv.py
@@ -323,6 +323,12 @@ def generate_no_sw_variant(net):
     for b1, b2 in to_fuse.items():
         pp.fuse_buses(net, b1, b2)
 
+    # replace auxiliary type of buses with no switch connected anymore
+    aux_buses = net.bus.index[net.bus.type == "auxiliary"]
+    buses_at_sw = net.switch.bus.append(net.switch.element.loc[net.switch.et == "b"])
+    aux_buses_to_change_type = aux_buses[~aux_buses.isin(buses_at_sw)]
+    net.bus.type.loc[aux_buses_to_change_type] = "b"
+
 
 def get_simbench_net(sb_code_info, input_path=None):
     """